pyproject.toml
[project]
name = "mcp-ragster"
version = "0.3.0"
description = "Model Context Protocol (MCP) server for RAG using Jina, Firecrawl, Perplexity, Milvus, and Voyage AI with improved error handling."
authors = [{ name = "Gemini", email = "gemini@google.com" }]
requires-python = ">=3.13, <3.14" 
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.110.0,<0.112.0",
    "uvicorn[standard]>=0.29.0,<0.31.0",
    "pymilvus>=2.4.0,<2.5.0",
    "firecrawl-py>=0.0.15,<0.1.0",
    "httpx>=0.27.0,<0.28.0",
    "python-dotenv>=1.0.0,<1.1.0",
    "aiohttp>=3.9.0,<3.10.0",
    "mcp[cli]>=1.9.1",
]

[project.scripts]
start-mcp-server = "ragster.main:main"

[tool.uv]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.401",
    "ruff>=0.11.11",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

